kazi-agent/python/requirements.txt
pyautogui>=0.9.54
pillow>=10.0.0
mss>=9.0.0
google-generativeai>=0.8.0
keyboard>=0.13.5
pyperclip>=1.8.2
cairosvg>=2.7.0
